:white_check_mark: Update: Some changes

+++ src/main/java/com/integrador/svfapi/dto/dashboardDTO/MonthPensionsCount.java
@@ -8,5 +8,5 @@ import lombok.Setter;
 @Data
 public class MonthPensionsCount {
     private int monthNumber;
-    private long count;
+    private int count;
 }
